Package structure: moved functionality from FileUtils to MarkerUtils.
Part of #963.
